// client/src/app/shared/user.ts
export class User {
    id: number;
    name: string;
    firstname: string;
    email: string;
    city: string;
    linkedin: string;
    github: string;
    resume: string;
    techskill: string;
    language: string;
    softskill: string;
    role: string;
    photo: string;
    password: string;
    
}